// src/impl/formats.d.ts
/**
 * @private
 */
export declare const DATE_SHORT: {
    year: string;
    month: string;
    day: string;
};
export declare const DATE_MED: {
    year: string;
    month: string;
    day: string;
};
export declare const DATE_MED_WITH_WEEKDAY: {
    year: string;
    month: string;
    day: string;
    weekday: string;
};
export declare const DATE_FULL: {
    year: string;
    month: string;
    day: string;
};
export declare const DATE_HUGE: {
    year: string;
    month: string;
    day: string;
    weekday: string;
};
export declare const TIME_SIMPLE: {
    hour: string;
    minute: string;
};
export declare const TIME_WITH_SECONDS: {
    hour: string;
    minute: string;
    second: string;
};
export declare const TIME_WITH_SHORT_OFFSET: {
    hour: string;
    minute: string;
    second: string;
    timeZoneName: string;
};
export declare const TIME_WITH_LONG_OFFSET: {
    hour: string;
    minute: string;
    second: string;
    timeZoneName: string;
};
export declare const TIME_24_SIMPLE: {
    hour: string;
    minute: string;
    hour12: boolean;
};
export declare const TIME_24_WITH_SECONDS: {
    hour: string;
    minute: string;
    second: string;
    hour12: boolean;
};
export declare const TIME_24_WITH_SHORT_OFFSET: {
    hour: string;
    minute: string;
    second: string;
    hour12: boolean;
    timeZoneName: string;
};
export declare const TIME_24_WITH_LONG_OFFSET: {
    hour: string;
    minute: string;
    second: string;
    hour12: boolean;
    timeZoneName: string;
};
export declare const DATETIME_SHORT: {
    year: string;
    month: string;
    day: string;
    hour: string;
    minute: string;
};
export declare const DATETIME_SHORT_WITH_SECONDS: {
    year: string;
    month: string;
    day: string;
    hour: string;
    minute: string;
    second: string;
};
export declare const DATETIME_MED: {
    year: string;
    month: string;
    day: string;
    hour: string;
    minute: string;
};
export declare const DATETIME_MED_WITH_SECONDS: {
    year: string;
    month: string;
    day: string;
    hour: string;
    minute: string;
    second: string;
};
export declare const DATETIME_MED_WITH_WEEKDAY: {
    year: string;
    month: string;
    day: string;
    weekday: string;
    hour: string;
    minute: string;
};
export declare const DATETIME_FULL: {
    year: string;
    month: string;
    day: string;
    hour: string;
    minute: string;
    timeZoneName: string;
};
export declare const DATETIME_FULL_WITH_SECONDS: {
    year: string;
    month: string;
    day: string;
    hour: string;
    minute: string;
    second: string;
    timeZoneName: string;
};
export declare const DATETIME_HUGE: {
    year: string;
    month: string;
    day: string;
    weekday: string;
    hour: string;
    minute: string;
    timeZoneName: string;
};
export declare const DATETIME_HUGE_WITH_SECONDS: {
    year: string;
    month: string;
    day: string;
    weekday: string;
    hour: string;
    minute: string;
    second: string;
    timeZoneName: string;
};